feat(events): prevent deletion of historical records based on state machine

+++ app/routers/events.py
@@ -152,6 +152,8 @@ def delete_event (event_id: int, db: Session = Depends(get_db)):
     event = db.query(Event).filter(Event.id == event_id).first()
     if event is None:
         raise HTTPException(status_code=404, detail="Event not found")
+    if event.status in [StatusEvent.FINISHED, StatusEvent.CANCELED]:
+        raise HTTPException(status_code=400, detail="Finished events or Canceled events cannot be deleted" )
     
     #deletar os dados
     db.delete(event)
